Extract `printStatementSequence`
