+++ src/language-js/print/statement.js
@@ -1,0 +1,139 @@
+"use strict";
+
+/** @typedef {import("../document").Doc} Doc */
+
+const { isNextLineEmpty } = require("../../common/util");
+const {
+  builders: { concat, join, hardline },
+} = require("../../document");
+const pathNeedsParens = require("../needs-parens");
+const {
+  classChildNeedsASIProtection,
+  classPropMayCauseASIProblems,
+  getLeftSidePathName,
+  hasNakedLeftSide,
+  isJSXNode,
+  isLastStatement,
+  isTheOnlyJSXElementInMarkdown,
+} = require("../utils");
+const { locEnd } = require("../loc");
+
+const { shouldPrintParamsWithoutParens } = require("./function");
+
+function printStatementSequence(path, options, print) {
+  const printed = [];
+
+  const bodyNode = path.getNode();
+  const isClass = bodyNode.type === "ClassBody";
+
+  path.each((stmtPath, i) => {
+    const stmt = stmtPath.getValue();
+
+    // Just in case the AST has been modified to contain falsy
+    // "statements," it's safer simply to skip them.
+    /* istanbul ignore if */
+    if (!stmt) {
+      return;
+    }
+
+    // Skip printing EmptyStatement nodes to avoid leaving stray
+    // semicolons lying around.
+    if (stmt.type === "EmptyStatement") {
+      return;
+    }
+
+    const stmtPrinted = print(stmtPath);
+    const text = options.originalText;
+    const parts = [];
+
+    // in no-semi mode, prepend statement with semicolon if it might break ASI
+    // don't prepend the only JSX element in a program with semicolon
+    if (
+      !options.semi &&
+      !isClass &&
+      !isTheOnlyJSXElementInMarkdown(options, stmtPath) &&
+      stmtNeedsASIProtection(stmtPath, options)
+    ) {
+      if (stmt.comments && stmt.comments.some((comment) => comment.leading)) {
+        parts.push(print(stmtPath, { needsSemi: true }));
+      } else {
+        parts.push(";", stmtPrinted);
+      }
+    } else {
+      parts.push(stmtPrinted);
+    }
+
+    if (!options.semi && isClass) {
+      if (classPropMayCauseASIProblems(stmtPath)) {
+        parts.push(";");
+      } else if (
+        stmt.type === "ClassProperty" ||
+        stmt.type === "FieldDefinition"
+      ) {
+        const nextChild = bodyNode.body[i + 1];
+        if (classChildNeedsASIProtection(nextChild)) {
+          parts.push(";");
+        }
+      }
+    }
+
+    if (isNextLineEmpty(text, stmt, locEnd) && !isLastStatement(stmtPath)) {
+      parts.push(hardline);
+    }
+
+    printed.push(concat(parts));
+  });
+
+  return join(hardline, printed);
+}
+
+function stmtNeedsASIProtection(path, options) {
+  const node = path.getNode();
+
+  if (node.type !== "ExpressionStatement") {
+    return false;
+  }
+
+  return path.call(
+    (childPath) => exprNeedsASIProtection(childPath, options),
+    "expression"
+  );
+}
+
+function exprNeedsASIProtection(path, options) {
+  const node = path.getValue();
+
+  const maybeASIProblem =
+    pathNeedsParens(path, options) ||
+    node.type === "ParenthesizedExpression" ||
+    node.type === "TypeCastExpression" ||
+    (node.type === "ArrowFunctionExpression" &&
+      !shouldPrintParamsWithoutParens(path, options)) ||
+    node.type === "ArrayExpression" ||
+    node.type === "ArrayPattern" ||
+    (node.type === "UnaryExpression" &&
+      node.prefix &&
+      (node.operator === "+" || node.operator === "-")) ||
+    node.type === "TemplateLiteral" ||
+    node.type === "TemplateElement" ||
+    isJSXNode(node) ||
+    (node.type === "BindExpression" && !node.object) ||
+    node.type === "RegExpLiteral" ||
+    (node.type === "Literal" && node.pattern) ||
+    (node.type === "Literal" && node.regex);
+
+  if (maybeASIProblem) {
+    return true;
+  }
+
+  if (!hasNakedLeftSide(node)) {
+    return false;
+  }
+
+  return path.call(
+    (childPath) => exprNeedsASIProtection(childPath, options),
+    ...getLeftSidePathName(path, node)
+  );
+}
+
+module.exports = { printStatementSequence };
